requirements.txt
# Requirements. These can can be installed with:
#   pip install -r requirements.txt

# You will need a C compiler and proper setup for Cython
# https://wiki.python.org/moin/WindowsCompilers


    aiohttp~=3.8.3
    aiohttp_cors~=0.7.0
    aubio>=0.4.9
    cython~=0.29.21
    certifi~=2020.12.5
    multidict~=5.0.0
    numpy~=1.26.2
    openrgb-python~=0.2.10
    paho-mqtt~=1.5.1
    psutil>=5.8.0
    pyserial~=3.5
    pystray>=0.17
    python-rtmidi~=1.5.6
    requests~=2.28.2
    sacn~=1.6.3
    sentry-sdk==1.14.0
    samplerate>=0.1.0
    sounddevice~=0.4.2
    icmplib~=3.0.3
    voluptuous~=0.12.1
    yappi~=1.4
    zeroconf~=0.39.4
    flux-led>=0.28.35
    python-mbedtls~=2.7.1; (platform_system == 'Linux' and platform_machine != 'aarch64' and platform_machine != 'armv7l') or platform_system == 'Windows' or platform_system == 'Darwin'
    python-osc~=1.8.3

